uploadApp/requirements.txt
Flask==1.0.2
google-cloud-storage==1.19.1
gunicorn==19.6.0
oauth2client==3.0.0